# LibGfx/PNGWriter: Compute which predictor to use first, store data then